requirements.txt
selenium
requests
reportlab
beautifulsoup4
python-dotenv
webdriver-manager
urllib3